// Frontend/src/app/types/user.ts
//TODO: Add username 
export type UserProfile = {
	userId: number;
	username: string;
	// age: number;
	// weight: number;
	// country: string;
	// avatar: string;
	// status: string;
	reputation: number;
	// funFact: string;
	covidVaccine: boolean;
	smoker: boolean;
	drinker: boolean;
	// twoFactor: boolean;
	optOutOfPublicStories: boolean;
	notificationPermission: boolean;
	// filePermission: boolean;
	// nightMode: boolean;
	// highContrast: boolean;
	// slowInternet: boolean;
	textSize: number;
	// screenReader: boolean;
};